Add full functionality when reading from MKV files (#6)
This enables the ability to extract and create the srt output for each
file in a give mkv input. These files will all be saved to the directory
provided as the second command line argument. This also means that when
operating on a simple sup input the output will be saved to the
directory given as the second command line argument. 

Signed-off-by: Ethan Dye <mrtops03@gmail.com>