Fix handling of single entry PDS segments (#16)
Previously the code required that there be more than one palette entry
in a single PDS segment. This fixes that bug to allow a single palette
entry to properly be parsed. Additionally it add slightly more robust
error handling to parts of the PGS code. This should be refined in the
future to be present throughout the codebase.

Signed-off-by: Ethan Dye <mrtops03@gmail.com>