Fix decoding of VobSub for different RLE offsets
This should fix any issues that could occur due to the data being offset
from the typical 0 in the RLE buffer. However this has not been tested
against a large number of subtitles, so there may be some cases where it
either breaks decoding or doesn't properly interact with various RLE
offsets.

Fixes #19

Signed-off-by: Ethan Dye <mrtops03@gmail.com>