Use FFmpeg as the default decoder (#22)
This will use FFmpeg as the default decoder for now. This will allow for
more accurate decoding of files, and perhaps better debugging as well.
This will also still keep the internal decoder, so that it can be worked
on until it has all of its bugs ironed out. In the meantime it has been
labled as experimental and is only used when requested.

---------

Signed-off-by: Ethan Dye <mrtops03@gmail.com>